% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_antrax_output}
\alias{read_antrax_output}
\alias{transform_antrax_output}
\title{Read the primary data}
\usage{
read_antrax_output(antrex_output_dir)

transform_antrax_output(data)
}
\arguments{
\item{antrex_output_dir}{}

\item{data}{}
}
\description{
Read the .csv files of tracking data
}
\section{Functions}{
\itemize{
\item \code{read_antrax_output}: read the output from anTrax tracking

\item \code{transform_antrax_output}: transform the raw data into r readable pivot tibble
}}

